+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 - Option to register route53 cnames for vpc endpoints - see [#20](https://github.com/ExpediaInc/apiary-waggledance/issues/20).
+- Option to associate multiple VPCs to Service Discovery namespace - see [#29](https://github.com/ExpediaInc/apiary-waggledance/issues/29)
 - Enable write support for federated metastores - see [#26](https://github.com/ExpediaInc/apiary-waggledance/issues/26).
 
 ### Removed